remove paranthesis from components directory
