+++ src/components/BlogCard/BlogCard.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/globalColor.scss";
+@import "../../styles/globalColor.scss";
 
 .container {
   border-radius: 10px;
