+++ src/containers/Blogs/Blogs.module.scss
@@ -1,5 +1,5 @@
 /*Blog Cards Layout*/
-@import "../../../styles/globalColor.scss";
+@import "../../styles/globalColor.scss";
 
 .headerText {
   font-size: 56px;
